+++ Plugins/Android/GamesServicesBridge.androidlib/src/main/java/com/bizsim/google/play/games/cloudsave/ICloudSaveCallback.java
@@ -4,10 +4,10 @@ package com.bizsim.google.play.games.cloudsave;
 
 public interface ICloudSaveCallback {
     void onSnapshotOpened(String filename, String snapshotJson, boolean hasConflict);
-    void onSnapshotRead(String filename, byte[] data);
+    void onSnapshotRead(String filename, String dataBase64);
     void onSnapshotCommitted(String filename);
     void onSnapshotDeleted(String filename);
     void onSavedGamesUIResult(String selectedFilename);
-    void onConflictDetected(String localSnapshotJson, String serverSnapshotJson, byte[] localData, byte[] serverData);
+    void onConflictDetected(String localSnapshotJson, String serverSnapshotJson, String localDataBase64, String serverDataBase64);
     void onCloudSaveError(int errorCode, String errorMessage, String filename);
 }